+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This changelog includes only user-sided updates. It indicates the advancement fo
 - ✔ Clean overall theme
 - ✔ End screen with useful information (score place, score total, restart option, best combo, number of clicks)
 - ✔️ Visual effects on touch
-- 🏳️ Animations
+- ✔️ Animations
 - 🏳 Sound
 - 🏳 Statistics on the scoreboard
+- 🏳 Settings to change animation speed, and game colors